requirements.txt
PyWavelets==1.7.0
pillow==10.4.0
numpy==2.1.1
scipy==1.14.1
fastapi[standard]==0.115.5